+++ src/EntityGraphQL/Schema/ISchemaType.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Linq.Expressions;
 
 namespace EntityGraphQL.Schema
 {
@@ -9,6 +10,7 @@ namespace EntityGraphQL.Schema
         string Name { get; }
         string Description { get; }
         bool IsInput { get; }
+        bool IsEnum { get; }
 
         Field GetField(string identifier);
         IEnumerable<Field> GetFields();
@@ -16,5 +18,10 @@ namespace EntityGraphQL.Schema
         void AddFields(List<Field> fields);
         void AddField(Field field);
         void RemoveField(string name);
+        /// <summary>
+        /// Add all fields and properties from the dotnet type to the schema type
+        /// </summary>
+        /// <returns></returns>
+        ISchemaType AddAllFields();
     }
 }